trustpulse/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pymysql==1.1.1
cryptography==42.0.7
pydantic==2.7.1
python-multipart==0.0.9
numpy==1.26.4
python-dateutil==2.9.0
aiofiles==23.2.1
jinja2==3.1.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.6
